fix: cast i64 comparison ops result to i32 to preserve Wasm op semantics

+++ frontend-wasm/src/code_translator/tests.rs
@@ -1159,7 +1159,7 @@ fn i64_lt_u() {
             v2 = cast v0  : u64
             v3 = cast v1  : u64
             v4 = lt v2, v3  : i1
-            v5 = cast v4  : i64
+            v5 = cast v4  : i32
         "#]],
     )
 }
@@ -1195,7 +1195,7 @@ fn i64_lt_s() {
             v0 = const.i64 2  : i64
             v1 = const.i64 1  : i64
             v2 = lt v0, v1  : i1
-            v3 = cast v2  : i64
+            v3 = cast v2  : i32
         "#]],
     )
 }
@@ -1235,7 +1235,7 @@ fn i64_le_u() {
             v2 = cast v0  : u64
             v3 = cast v1  : u64
             v4 = lte v2, v3  : i1
-            v5 = cast v4  : i64
+            v5 = cast v4  : i32
         "#]],
     )
 }
@@ -1271,7 +1271,7 @@ fn i64_le_s() {
             v0 = const.i64 2  : i64
             v1 = const.i64 1  : i64
             v2 = lte v0, v1  : i1
-            v3 = cast v2  : i64
+            v3 = cast v2  : i32
         "#]],
     )
 }
@@ -1311,7 +1311,7 @@ fn i64_gt_u() {
             v2 = cast v0  : u64
             v3 = cast v1  : u64
             v4 = gt v2, v3  : i1
-            v5 = cast v4  : i64
+            v5 = cast v4  : i32
         "#]],
     )
 }
@@ -1387,7 +1387,7 @@ fn i64_ge_u() {
             v2 = cast v0  : u64
             v3 = cast v1  : u64
             v4 = gte v2, v3  : i1
-            v5 = cast v4  : i64
+            v5 = cast v4  : i32
         "#]],
     )
 }
@@ -1423,7 +1423,7 @@ fn i64_ge_s() {
             v0 = const.i64 2  : i64
             v1 = const.i64 1  : i64
             v2 = gte v0, v1  : i1
-            v3 = cast v2  : i64
+            v3 = cast v2  : i32
         "#]],
     )
 }
@@ -1457,7 +1457,7 @@ fn i64_eqz() {
             v0 = const.i64 2  : i64
             v1 = const.i64 0  : i64
             v2 = eq v0, v1  : i1
-            v3 = cast v2  : i64
+            v3 = cast v2  : i32
         "#]],
     )
 }
@@ -1493,7 +1493,7 @@ fn i64_eq() {
             v0 = const.i64 2  : i64
             v1 = const.i64 1  : i64
             v2 = eq v0, v1  : i1
-            v3 = cast v2  : i64
+            v3 = cast v2  : i32
         "#]],
     )
 }
@@ -1547,7 +1547,7 @@ fn i64_ne() {
             v0 = const.i64 2  : i64
             v1 = const.i64 1  : i64
             v2 = neq v0, v1  : i1
-            v3 = cast v2  : i64
+            v3 = cast v2  : i32
         "#]],
     )
 }
